cmake: restrict GGML_CPU_ALL_VARIANTS to supported architectures
The previous logic for enabling GGML_CPU_ALL_VARIANTS used a blacklist to exclude ARM processors. This caused builds to fail on other architectures not in the blacklist and not compatible with the option, such as loongarch64.

This changes the condition to a whitelist, only enabling GGML_CPU_ALL_VARIANTS for x86 and PowerPC architectures.

Signed-off-by: Zhou Qiankang <wszqkzqk@qq.com>